feat(controller): DTO 매핑 객체 분리
