+++ src/main/java/controller/BlackJackController.java
@@ -1,26 +1,14 @@
 package controller;
 
-import domain.card.Card;
+import controller.mapper.ResultViewMapper;
 import domain.game.BlackJackGame;
 import domain.card.Cards;
 import domain.participant.Dealer;
 import domain.game.GameResult;
 import domain.participant.Player;
 import domain.participant.Players;
-import domain.state.Outcome;
 import dto.domain.PlayerNameAndBettingDto;
-import dto.view.DealerStatDto;
-import dto.view.ParticipantProfitDto;
-import dto.view.ParticipantResultDto;
-import dto.view.ParticipantStatsDto;
-import dto.view.PlayerCardsDto;
-import dto.view.PlayerOutcomeDto;
-import dto.view.PlayerProfitDto;
-import dto.view.ResultDto;
-import dto.view.StartBlackJackDto;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 import util.NameParser;
 import view.InputView;
 import view.ResultView;
@@ -28,10 +16,12 @@ import view.ResultView;
 public class BlackJackController {
     private final InputView inputView;
     private final ResultView resultView;
+    private final ResultViewMapper resultViewMapper;
 
     public BlackJackController(InputView inputView, ResultView resultView) {
         this.inputView = inputView;
         this.resultView = resultView;
+        this.resultViewMapper = new ResultViewMapper(resultView);
     }
 
     public void run() {
@@ -42,7 +32,7 @@ public class BlackJackController {
         final Dealer dealer = game.dealer();
         final Cards cards = game.cards();
 
-        resultView.printGameStartMessage(toStartBlackJackDto(players, dealer));
+        resultView.printGameStartMessage(resultViewMapper.toStartBlackJackDto(players, dealer));
         drawPlayersTurn(players, cards);
         drawDealerTurn(dealer, cards);
         printResult(game);
@@ -88,91 +78,14 @@ public class BlackJackController {
         final Dealer dealer = game.dealer();
         final GameResult gameResult = game.calculateResult();
         printGameResult(players, dealer);
-        resultView.printWinner(toParticipantStatsDto(players, gameResult));
-        resultView.printFinalProfit(toParticipantProfitDto(players));
+        resultView.printWinner(resultViewMapper.toParticipantStatsDto(players, gameResult));
+        resultView.printFinalProfit(resultViewMapper.toParticipantProfitDto(players));
     }
 
     private void printGameResult(Players players, Dealer dealer) {
-        resultView.printResult(toResultDto(players, dealer));
+        resultView.printResult(resultViewMapper.toResultDto(players, dealer));
         if (dealer.checkBust()) {
             resultView.printDealerBust();
         }
-    }
-
-    private StartBlackJackDto toStartBlackJackDto(Players players, Dealer dealer) {
-        return new StartBlackJackDto(
-                formatCardName(dealer.getCardList().get(0)),
-                toPlayerCardsDtos(players)
-        );
-    }
-
-    private List<PlayerCardsDto> toPlayerCardsDtos(Players players) {
-        final List<PlayerCardsDto> playerCardsDtos = new ArrayList<>();
-        players.forEachPlayer(player -> playerCardsDtos.add(new PlayerCardsDto(
-                player.getName(),
-                toCardNames(player.getCardList())
-        )));
-        return playerCardsDtos;
-    }
-
-    private ResultDto toResultDto(Players players, Dealer dealer) {
-        final ParticipantResultDto dealerResult = new ParticipantResultDto(
-                "딜러",
-                resultView.joinCardNames(dealer.getCardList()),
-                dealer.getScore()
-        );
-        final List<ParticipantResultDto> playerResults = new ArrayList<>();
-        players.forEachPlayer(player -> playerResults.add(new ParticipantResultDto(
-                player.getName(),
-                resultView.joinCardNames(player.getCardList()),
-                player.getScore()
-        )));
-        return new ResultDto(dealerResult, playerResults);
-    }
-
-    private ParticipantStatsDto toParticipantStatsDto(Players players, GameResult gameResult) {
-        return new ParticipantStatsDto(
-                toDealerStatDto(gameResult),
-                toPlayerOutcomeDtos(players, gameResult)
-        );
-    }
-
-    private DealerStatDto toDealerStatDto(GameResult gameResult) {
-        final int dealerWinCount = gameResult.getDealerCount(Outcome.DEFAULT_WIN)
-                + gameResult.getDealerCount(Outcome.BLACKJACK_WIN);
-        final int dealerLoseCount = gameResult.getDealerCount(Outcome.LOSE);
-        final int dealerDrawCount = gameResult.getDealerCount(Outcome.DRAW);
-        return new DealerStatDto(dealerWinCount, dealerLoseCount, dealerDrawCount);
-    }
-
-    private List<PlayerOutcomeDto> toPlayerOutcomeDtos(Players players, GameResult gameResult) {
-        final List<PlayerOutcomeDto> playerOutcomeDtos = new ArrayList<>();
-        players.forEachPlayer(player -> playerOutcomeDtos.add(new PlayerOutcomeDto(
-                player.getName(),
-                gameResult.getPlayerOutcome(player.getName()).result()
-        )));
-        return playerOutcomeDtos;
-    }
-
-    private ParticipantProfitDto toParticipantProfitDto(Players players) {
-        final List<PlayerProfitDto> playerProfitDtos = new ArrayList<>();
-        players.forEachPlayer(player -> playerProfitDtos.add(new PlayerProfitDto(
-                player.getName(),
-                player.getBalance()
-        )));
-        final int dealerProfit = -playerProfitDtos.stream()
-                .mapToInt(PlayerProfitDto::profit)
-                .sum();
-        return new ParticipantProfitDto(dealerProfit, playerProfitDtos);
-    }
-
-    private List<String> toCardNames(List<Card> cards) {
-        return cards.stream()
-                .map(this::formatCardName)
-                .collect(Collectors.toList());
-    }
-
-    private String formatCardName(Card card) {
-        return card.getRank().symbol() + card.getSuit().suit();
     }
 }
