+++ src/test/java/domain/game/BlackJackGameTest.java
@@ -2,6 +2,8 @@ package domain.game;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import dto.domain.PlayerNameAndBettingDto;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -9,7 +11,10 @@ public class BlackJackGameTest {
     @Test
     @DisplayName("게임 시작 시 딜러와 모든 플레이어는 카드 2장을 가진다")
     void startGameWithInitialCards() {
-        BlackJackGame game = BlackJackGame.startGame("pobi,jason");
+        BlackJackGame game = BlackJackGame.startGame(List.of(
+                new PlayerNameAndBettingDto("pobi", 1000),
+                new PlayerNameAndBettingDto("jason", 1000)
+        ));
 
         assertEquals(2, game.dealer().getCardList().size());
         game.players().forEachPlayer(player -> assertEquals(2, player.getCardList().size()));
