+++ src/test/java/domain/participant/PlayerTest.java
@@ -11,7 +11,7 @@ public class PlayerTest {
     @Test
     @DisplayName("플레이어가 카드를 받으면 카드 개수가 1 증가한다")
     void drawCardByPlayer() {
-        Player player = new Player("pobi");
+        Player player = new Player("pobi", 1000);
         Cards cards = new Cards(new Random(1));
 
         player.drawCard(cards);
