+++ src/test/java/domain/participant/PlayersTest.java
@@ -11,7 +11,7 @@ public class PlayersTest {
     @Test
     @DisplayName("Players는 생성 시 전달된 플레이어 수를 보관한다")
     void playersSize() {
-        Players players = new Players(List.of(new Player("pobi"), new Player("jason")));
+        Players players = new Players(List.of(new Player("pobi", 1000), new Player("jason", 1000)));
 
         assertEquals(2, players.getSize());
     }
@@ -19,7 +19,7 @@ public class PlayersTest {
     @Test
     @DisplayName("forEachPlayer로 플레이어 이름 순회를 할 수 있다")
     void forEachPlayer() {
-        Players players = new Players(List.of(new Player("pobi"), new Player("jason")));
+        Players players = new Players(List.of(new Player("pobi", 1000), new Player("jason", 1000)));
         List<String> names = new ArrayList<>();
 
         players.forEachPlayer(player -> names.add(player.getName()));
